… src/Fantomas.Benchmarks/packages.lock.json
… src/Fantomas.Core.Tests/packages.lock.json
… src/Fantomas.Core/Context.fs
… src/Fantomas.Core/ImmutableArray.fsi
… src/Fantomas.Core/Selection.fs
… src/Fantomas.Core/SyntaxOak.fs
… src/Fantomas.Core/Trivia.fs
… src/Fantomas.Core/packages.lock.json
… src/Fantomas.Tests/packages.lock.json
… src/Fantomas/packages.lock.json